user story #623123 : UFT integration - rename test in the repository should rename the test in Octane
