+++ pom.xml
@@ -465,7 +465,7 @@
 		<dependency>
 			<artifactId>mqm-rest-client</artifactId>
 			<groupId>com.hpe.adm.octane.ciplugins</groupId>
-			<version>1.4.23</version>
+			<version>1.4.25</version>
 		</dependency>
 
 		<!-- org.jenkins-ci.main -->
